Accidently changed the encoder to double so changed it back

+++ src/main/java/frc/robot/Constants.java
@@ -101,7 +101,7 @@ public final class Constants {
         public static final int ShintakePivot_FOLLOW_MOTOR = 43;
         public static final double ShintakePivot_MAX_ANGLE = 2;
         public static final double ShintakePivot_MIN_ANGLE = 3;
-        public static final double ShintakePivot_ENCODER = 4;
+        public static final int ShintakePivot_ENCODER = 4;
         public static final double ShintakePivot_GEAR_RATIO = (40.0 / 11.0) * (20.0 / 1.0);
         public static final double ShintakePivot_ENCODER_OFFSET = 6;
     }
